+++ src/main/java/blackjack/domain/user/exceptions/MoneyException.java
@@ -1,4 +1,4 @@
-package blackjack.domain.betting.exceptions;
+package blackjack.domain.user.exceptions;
 
 public class MoneyException extends RuntimeException {
 	public MoneyException(String message) {
