+++ src/main/java/blackjack/domain/yesOrNo/exceptions/YesOrNoException.java
@@ -1,4 +1,4 @@
-package blackjack.domain.user.exceptions;
+package blackjack.domain.yesOrNo.exceptions;
 
 public class YesOrNoException extends RuntimeException {
 	public YesOrNoException(String message) {
